test(integration): update cypress
We can now run cypress on our Apple Silicon Macs!
