+++ integration-testing/bsd/cypress.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'cypress';
+
+/**
+ * Define the configuration for the Cypress tests.
+ */
+export default defineConfig({
+  viewportHeight: 1080,
+  viewportWidth: 1920,
+  e2e: {
+    baseUrl: 'http://localhost:3000',
+  },
+});
