--- integration-testing/bsd/cypress.json
@@ -1,5 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000",
-  "viewportHeight": 1080,
-  "viewportWidth": 1920
-}
