+++ integration-testing/bsd/cypress/e2e/configuration.cy.ts
