+++ integration-testing/bsd/package.json
@@ -38,16 +38,17 @@
     ]
   },
   "dependencies": {
+    "@testing-library/cypress": "^8.0.3",
+    "@votingworks/fixtures": "workspace:*",
     "buffer": "^6.0.3",
-    "js-sha256": "^0.9.0"
+    "cypress": "^10.3.1",
+    "cypress-file-upload": "^5.0.7",
+    "js-sha256": "^0.9.0",
+    "start-server-and-test": "^1.12.5"
   },
   "devDependencies": {
-    "@testing-library/cypress": "^8.0.2",
     "@typescript-eslint/eslint-plugin": "^5.21.0",
     "@typescript-eslint/parser": "^5.21.0",
-    "@votingworks/fixtures": "workspace:*",
-    "cypress": "^9.5.4",
-    "cypress-file-upload": "^5.0.7",
     "eslint": "^7.17.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
@@ -64,7 +65,7 @@
     "lint-staged": "^10.5.4",
     "prettier": "^2.6.2",
     "sort-package-json": "^1.50.0",
-    "start-server-and-test": "^1.12.5",
     "typescript": "4.6.3"
-  }
+  },
+  "packageManager": "pnpm@5.18.10"
 }
