+++ integration-testing/election-manager/cypress/e2e/candidate_contest_tallies.cy.ts
