+++ integration-testing/election-manager/cypress/e2e/configuration.cy.ts
