+++ integration-testing/election-manager/cypress/e2e/tallies.cy.ts
