+++ integration-testing/election-manager/cypress/e2e/yes_no_contest_tallies.cy.ts
