+++ integration-testing/election-manager/cypress/support/e2e.ts
