+++ integration-testing/election-manager/package.json
@@ -37,15 +37,19 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {
-    "@testing-library/cypress": "^8.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
+  "dependencies": {
+    "@testing-library/cypress": "^8.0.3",
     "@votingworks/fixtures": "workspace:*",
     "@votingworks/test-utils": "workspace:*",
     "@votingworks/types": "workspace:*",
-    "cypress": "^9.5.4",
+    "cypress": "^10.3.1",
     "cypress-file-upload": "^5.0.7",
+    "start-server-and-test": "^1.12.5"
+  },
+  "devDependencies": {
+    "@types/node": "^18.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.21.0",
+    "@typescript-eslint/parser": "^5.21.0",
     "eslint": "^7.17.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
@@ -61,7 +65,7 @@
     "lint-staged": "^11.0.0",
     "prettier": "^2.6.2",
     "sort-package-json": "^1.50.0",
-    "start-server-and-test": "^1.12.5",
     "typescript": "4.6.3"
-  }
+  },
+  "packageManager": "pnpm@5.18.10"
 }
