+++ integration-testing/election-manager/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tsconfig.shared.json",
+  "extends": "../../tsconfig.shared.json",
   "compilerOptions": {
     "types": ["cypress", "cypress-file-upload"],
     "isolatedModules": false,
@@ -18,9 +18,9 @@
   },
   "include": ["**/*.ts"],
   "references": [
-    { "path": "../../../libs/eslint-plugin-vx/tsconfig.build.json" },
-    { "path": "../../../libs/fixtures/tsconfig.build.json" },
-    { "path": "../../../libs/test-utils/tsconfig.build.json" },
-    { "path": "../../../libs/types/tsconfig.build.json" }
+    { "path": "../../libs/eslint-plugin-vx/tsconfig.build.json" },
+    { "path": "../../libs/fixtures/tsconfig.build.json" },
+    { "path": "../../libs/test-utils/tsconfig.build.json" },
+    { "path": "../../libs/types/tsconfig.build.json" }
   ]
 }
